+++ public/Images/index.tsx
@@ -1,4 +1,4 @@
-import HeaderBg from "./HeaderBg.svg";
+import HeaderBg from "./Top_colors.png";
 import FooterBg from "./FooterBg.svg";
 
 const Images = {HeaderBg, FooterBg};
