+++ src/Icons/index.tsx
@@ -6,6 +6,10 @@ import Calendar from "./Calendar.svg";
 import Search from "./Search.svg";
 import Close from "./Close.png";
 
+/*
+  * Icons component
+  * @returns {JSX.Element} - A collection of SVG icons used in the application.
+*/
 const Icons = {Logo, BlinkingStar, ScanBarcode, Clock, Calendar, Search, Close};
 
 export default Icons;
