+++ src/components/Common/Buttons/Button/index.d.ts
@@ -1,0 +1,9 @@
+type ButtonProps = {
+    children: React.ReactNode | string;
+    className?: string;
+    onClick?: () => void;
+    disabled?: boolean;
+    type?: "button" | "submit" | "reset";
+};
+
+export default ButtonProps;
