+++ src/components/Common/Buttons/OutlinedButton/index.d.ts
@@ -1,0 +1,9 @@
+type OutlinedButtonProps = {
+    children: React.ReactNode | string;
+    onClick?: () => void;
+    className?: string;
+    disabled?: boolean;
+    type?: "button" | "submit" | "reset";
+};
+
+export default OutlinedButtonProps;
