+++ src/components/Common/Image/index.d.ts
@@ -1,0 +1,9 @@
+type ImageProps = {
+    src: string;
+    alt?: string;
+    width?: number;
+    height?: number;
+    className?: string;
+};
+
+export default ImageProps;
