+++ src/components/Common/VerticalSeperator/index.d.ts
@@ -1,0 +1,6 @@
+type VerticalSeperatorProps = {
+    height?: string;
+    color?: string;
+};
+
+export default VerticalSeperatorProps;
