+++ src/components/Common/VerticalSeperator/index.styles.tsx
