+++ src/components/Search/index.tsx
@@ -1,12 +1,16 @@
-import Button from "../Buttons/Button";
+import Button from "../Common/Buttons/Button";
 import DatePicker from "../DatePicker";
 import SearchableSelect from "../SearchableSelect";
 import TimePicker from "../TimePicker";
-import VerticalSeperator from "../VerticalSeperator";
+import VerticalSeperator from "../Common/VerticalSeperator";
 
 //css
 import { SearchWrapper, Wrapper } from "./index.styles";
 
+/*
+  * Search component
+  * @returns {JSX.Element} - A search component with date, time, and location pickers.
+*/
 const Search = () => {
   return (
     <SearchWrapper>
